Fix Motor's hostname-validation tests, which weren't running.
